[FRBB-25] Consilidate External JS
Kinda accomplishes the goals of this ticket: https://trello.com/c/VY7DTIf3/25-cleanup-code-consolidate-existing-js-into-external-js-files
but instead of a single JS, it's still inline JS, since we need to get the urls from django still. We are on a path of consolidation, but more importantly the JS is significantly easier to read and navigate about! I also understand a lot more about how the front end works, and have some new ideas about how we can consolidate.

Next up we should focus on styling the form buttons to appear properly.